requirements.txt
langchain
langchain-community
langchain-openai
langgraph
chromadb
openai
ollama
python-dotenv
PyPDF2
beautifulsoup4
